conda-requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
alabaster=0.7.12=py_0
altgraph=0.17=pyhd8ed1ab_0
argon2-cffi=20.1.0=py39h2bbff1b_1
astroid=2.5.8=py39hcbf5309_0
async_generator=1.10=pyhd3eb1b0_0
attrs=21.2.0=pyhd3eb1b0_0
babel=2.9.1=pyh44b312d_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
bottleneck=1.3.2=py39h7cc1a96_1
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py39hb82d6ee_1001
ca-certificates=2021.5.30=h5b45459_0
certifi=2021.5.30=py39hcbf5309_0
cffi=1.14.6=py39h2bbff1b_0
chardet=3.0.4=pypi_0
charset-normalizer=2.0.0=pyhd8ed1ab_0
click=8.0.1=pypi_0
colorama=0.4.4=pyhd3eb1b0_0
cryptography=3.4.7=py39hd8d06c1_0
cycler=0.10.0=py39haa95532_0
decorator=5.0.9=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
docutils=0.17.1=py39hcbf5309_0
easyocr=1.4=pypi_0
entrypoints=0.3=py39haa95532_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
future=0.18.2=py39hcbf5309_3
googletrans=3.0.0=pypi_0
h11=0.9.0=pypi_0
h2=3.2.0=pypi_0
hpack=3.0.0=pypi_0
hstspreload=2021.9.1=pypi_0
httpcore=0.9.1=pypi_0
httpx=0.13.3=pypi_0
hyperframe=5.2.0=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=pypi_0
imageio=2.9.0=pypi_0
imagesize=1.2.0=py_0
importlib-metadata=3.10.0=py39haa95532_0
importlib_metadata=3.10.0=hd3eb1b0_0
intel-openmp=2021.3.0=haa95532_3372
ipykernel=5.3.4=py39h7b7c402_0
ipython=7.26.0=py39hd4e2768_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pyhd3eb1b0_1
isort=5.9.3=pyhd8ed1ab_0
jedi=0.18.0=py39haa95532_1
jinja2=3.0.1=pyhd3eb1b0_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=py_2
jupyter=1.0.0=py39haa95532_7
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.7.1=py39haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.1=py39hd77b12b_0
lazy-object-proxy=1.6.0=py39hb82d6ee_0
libpng=1.6.37=h2a8f88b_0
libretranslatepy=2.1.1=pypi_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.2.0=hd0e1b90_0
lxml=4.6.3=pypi_0
lz4-c=1.9.3=h2bbff1b_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
macholib=1.14=pyh9f0ad1d_1
markupsafe=2.0.1=py39h2bbff1b_0
matplotlib=3.4.2=py39haa95532_0
matplotlib-base=3.4.2=py39h49ac443_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mccabe=0.6.1=py_1
mistune=0.8.4=py39h2bbff1b_1000
mkl=2021.3.0=haa95532_524
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.0=py39h277e83a_2
mkl_random=1.2.2=py39hf11a4ad_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=py_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py39haa95532_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.6.2=pypi_0
notebook=6.4.1=py39haa95532_0
numexpr=2.7.3=py39hb80d3ca_1
numpy=1.20.3=py39ha4e8547_0
numpy-base=1.20.3=py39hc2deb75_0
olefile=0.46=py_0
opencv-python=4.5.3.56=pypi_0
openssl=1.1.1k=h8ffe710_1
packaging=21.0=pyhd3eb1b0_0
pandas=1.3.1=py39h6214cd6_0
pandocfilters=1.4.3=py39haa95532_1
parso=0.8.2=pyhd3eb1b0_0
pefile=2021.5.24=pyhd8ed1ab_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.1=py39h4fa10fc_0
pip=21.2.2=py39haa95532_0
prometheus_client=0.11.0=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
prompt_toolkit=3.0.17=hd3eb1b0_0
pycparser=2.20=py_2
pygments=2.9.0=pyhd3eb1b0_0
pyinstaller=4.5.1=py39h4dafc3f_0
pyinstaller-hooks-contrib=2021.3=pyhd8ed1ab_0
pylint=2.7.2=py39hcbf5309_0
pynput=1.7.1=py39hde42818_1
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py39hd77b12b_6
pyrsistent=0.18.0=py39h2bbff1b_0
pysocks=1.7.1=py39hcbf5309_3
pytesseract=0.3.8=pyhd8ed1ab_0
python=3.9.6=h6244533_0
python-bidi=0.4.2=pypi_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-xlib=0.28=pyh9f0ad1d_0
python_abi=3.9=2_cp39
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=pypi_0
pywin32=228=py39hbaba5e8_1
pywin32-ctypes=0.2.0=py39hcbf5309_1003
pywinpty=0.5.7=py39haa95532_0
pyyaml=5.4.1=pypi_0
pyzmq=20.0.0=py39hd77b12b_1
qt=5.9.7=vc14h73c81de_0
qtconsole=5.1.0=pyhd3eb1b0_0
qtpy=1.9.0=py_0
rapidfuzz=1.5.1=pypi_0
requests=2.26.0=pyhd8ed1ab_0
rfc3986=1.5.0=pypi_0
scikit-image=0.18.2=pypi_0
scikit-learn=0.24.2=py39hf11a4ad_1
scipy=1.6.2=py39h66253e8_1
screen-ocr=0.2.0=pypi_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py39haa95532_0
setuptools-lint=0.6.0=pyh9f0ad1d_0
sip=4.19.13=py39hd77b12b_0
six=1.16.0=pyhd3eb1b0_0
sniffio=1.2.0=pypi_0
snowballstemmer=2.1.0=pyhd8ed1ab_0
sphinx=4.1.2=pyh6c4a22f_1
sphinxcontrib-applehelp=1.0.2=py_0
sphinxcontrib-devhelp=1.0.2=py_0
sphinxcontrib-htmlhelp=2.0.0=pyhd8ed1ab_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.3=py_0
sphinxcontrib-serializinghtml=1.1.5=pyhd8ed1ab_0
sqlite=3.36.0=h2bbff1b_0
terminado=0.9.4=py39haa95532_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyhb85f177_0
tifffile=2021.8.8=pypi_0
tk=8.6.10=he774522_0
toml=0.10.2=pyhd8ed1ab_0
torch=1.9.0=pypi_0
torchvision=0.10.0=pypi_0
tornado=6.1=py39h2bbff1b_0
traitlets=5.0.5=pyhd3eb1b0_0
translate=3.6.1=pypi_0
typing-extensions=3.10.0.0=pypi_0
tzdata=2021a=h52ac0ba_0
urllib3=1.26.6=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=py_0
webencodings=0.5.1=py39haa95532_1
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py39haa95532_0
win_inet_pton=1.1.0=py39hcbf5309_2
wincertstore=0.2=py39h2bbff1b_0
winpty=0.4.3=4
winrt=1.0.21033.1=pypi_0
wrapt=1.11.2=py39h4cdbadb_1
xz=5.2.5=h62dcd97_0
zeromq=4.3.3=ha925a31_3
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
